fix(frontend): resolve 65 TypeScript build errors
Root causes and fixes:

1. Missing sass compiler (Part 5.1 regression):
   - Installed sass@^1.70 (dart-sass) - sass-loader had no compiler
   - Upgraded tailwindcss 2.x postcss7-compat -> 3.4
   - Updated postcss 7 -> 8.4, autoprefixer 8 -> 10.4
   - Updated tailwind.config.js to v3 API (purge -> content)

2. My import removals in Part 3.4:
   - CommentItem.vue: restored defineComponent, reactive, toRefs, provide
   - Navigation.vue: restored defineComponent, reactive, toRefs
   - Social.vue: removed import (Options API, no Vue imports needed)

3. TypeScript strict null checks (30+ errors):
   - Added ! non-null assertions for appStore.websiteConfig
   - Added ! for userStore.userInfo and articleStore.topArticle
   - Changed userInfo === '' comparisons to === null

4. Type mismatches (15 errors):
   - commentStore.type: string -> number (used with numeric switch case)
   - Controls.vue: URLSearchParams typed as any
   - Controls.vue: articleId/password -> AccessArticleParams types
   - Controls.vue: accessArticles.push with Number() conversion
   - OauthLogin.vue: qqlogin params typed as any
   - Photos.vue: albumId Number() conversion
   - Talk.vue: talkId Number() conversion
   - Archives.vue: archives = '' -> archives = []

5. request.ts: InternalAxiosRequestConfig doesn't exist in axios 0.27.x

Result: npm run build passes with 0 errors.

+++ aurora-springboot/src/main/java/com/aurora/mapper/JobMapper.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface JobMapper extends BaseMapper<Job> {
 
-    Integer countJobs(@Param("jobSearchVO") JobSearchVO jobSearchVO);
+    Long countJobs(@Param("jobSearchVO") JobSearchVO jobSearchVO);
 
     List<JobDTO> listJobs(@Param("current") Long current, @Param("size") Long size, @Param("jobSearchVO")JobSearchVO jobSearchVO);
